Rename slideshow to checklist
