+++ app/src/main/java/com/example/volcanoseason3/MainActivity.kt
@@ -38,7 +38,7 @@ class MainActivity : AppCompatActivity() {
         appBarConfiguration = AppBarConfiguration(
             setOf(
 //                R.id.nav_home, R.id.nav_gallery, R.id.nav_slideshow
-                R.id.nav_home, R.id.nav_slideshow
+                R.id.nav_home, R.id.nav_checklist
             ), drawerLayout
         )
         setupActionBarWithNavController(navController, appBarConfiguration)
