+++ app/src/main/java/com/example/volcanoseason3/ui/checklist/ChecklistViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.volcanoseason3.ui.slideshow
+package com.example.volcanoseason3.ui.checklist
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class SlideshowViewModel : ViewModel() {
+class ChecklistViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is slideshow Fragment"
+        value = "This is checklist Fragment"
     }
     val text: LiveData<String> = _text
 }
