+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -9,8 +9,8 @@
             android:icon="@drawable/ic_menu_camera"
             android:title="@string/menu_home" />
         <item
-            android:id="@+id/nav_slideshow"
-            android:icon="@drawable/ic_menu_slideshow"
-            android:title="@string/menu_slideshow" />
+            android:id="@+id/nav_checklist"
+            android:icon="@drawable/ic_menu_checklist"
+            android:title="@string/menu_checklist" />
     </group>
 </menu>
