+++ app/src/main/res/values/strings.xml
@@ -8,5 +8,5 @@
     <string name="action_settings">Settings</string>
 
     <string name="menu_home">Home</string>
-    <string name="menu_slideshow">Slideshow</string>
+    <string name="menu_checklist">Checklist</string>
 </resources>
